# apt/portal-rest/api/main.py
# generated by fastapi-codegen:
#   filename:  openapi.json
#   timestamp: 2025-08-04T14:17:45+00:00

from __future__ import annotations

from typing import Optional, Union
from uuid import UUID

from fastapi import FastAPI, Header

from .models import (
    AggregateNode,
    Error,
    Experiment,
    ExperimentCreate,
    ExperimentList,
    ExperimentModify,
    ExtensionRequest,
    ManifestArray,
    PortalID,
    Profile,
    ProfileCreate,
    ProfileList,
    ProfileModify,
    ResGroup,
    ResGroupError,
    ResGroupList,
    ResGroupReservation,
    ResGroupSearch,
    ResGroupSearchResult,
    SnapshotRequest,
    SnapshotStatus,
    Token,
    Version,
)

app = FastAPI(
    title='Cloudlab Portal API',
    description='This document describes the Cloudlab Portal API.',
    version='1.0.0',
    license={
        'name': 'Apache 2.0',
        'identifier': 'Apache-2.0',
        'url': 'http://www.apache.org/licenses/LICENSE-2.0.txt',
    },
    servers=[{'url': 'http://boss.emulab.net:43795'}],
)


@app.get(
    '/experiments',
    response_model=ExperimentList,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
    },
    tags=['Experiments'],
)
def list_experiments(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    x__api__elaborate: Optional[str] = Header(None, alias='X-Api-Elaborate'),
    experiment_id: Optional[UUID] = None,
    creator: Optional[str] = None,
    project: Optional[str] = None,
) -> Union[ExperimentList, Error]:
    pass


@app.post(
    '/experiments',
    response_model=None,
    responses={
        '201': {'model': Experiment},
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '429': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def create_experiment(
    x__api__token: str = Header(..., alias='X-Api-Token'), body: ExperimentCreate = None
) -> Optional[Union[Experiment, Error]]:
    pass


@app.get(
    '/experiments/{experiment_id}',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def get_experiment(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    x__api__elaborate: Optional[str] = Header(None, alias='X-Api-Elaborate'),
    experiment_id: PortalID = ...,
) -> Union[Experiment, Error]:
    pass


@app.put(
    '/experiments/{experiment_id}',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def extend_experiment(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    body: ExtensionRequest = None,
) -> Union[Experiment, Error]:
    pass


@app.patch(
    '/experiments/{experiment_id}',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def modify_experiment(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    body: ExperimentModify = None,
) -> Union[Experiment, Error]:
    pass


@app.delete(
    '/experiments/{experiment_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def terminate_experiment(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Optional[Error]:
    pass


@app.get(
    '/experiments/{experiment_id}/manifests',
    response_model=ManifestArray,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def get_experiment_manifests(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Union[ManifestArray, Error]:
    pass


@app.get(
    '/experiments/{experiment_id}/node/{client_id}',
    response_model=AggregateNode,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def get_experiment_node(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
) -> Union[AggregateNode, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/node/{client_id}/powercycle',
    response_model=AggregateNode,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def powercycle_experiment_node(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
) -> Union[AggregateNode, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/node/{client_id}/reboot',
    response_model=AggregateNode,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def reboot_experiment_node(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
) -> Union[AggregateNode, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/node/{client_id}/reload',
    response_model=AggregateNode,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def reload_experiment_node(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
) -> Union[AggregateNode, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/node/{client_id}/start',
    response_model=AggregateNode,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def start_experiment_node(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
) -> Union[AggregateNode, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/node/{client_id}/stop',
    response_model=AggregateNode,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def stop_experiment_node(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
) -> Union[AggregateNode, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/nodes/powercycle',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def powercycle_experiment_nodes(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Union[Experiment, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/nodes/reboot',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def reboot_experiment_nodes(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Union[Experiment, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/nodes/reload',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def reload_experiment_nodes(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Union[Experiment, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/nodes/start',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def start_experiment_nodes(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Union[Experiment, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/nodes/stop',
    response_model=Experiment,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def stop_experiment_nodes(
    x__api__token: str = Header(..., alias='X-Api-Token'), experiment_id: PortalID = ...
) -> Union[Experiment, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/snapshot/{client_id}',
    response_model=SnapshotStatus,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def start_experiment_snapshot(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    client_id: str = ...,
    body: SnapshotRequest = None,
) -> Union[SnapshotStatus, Error]:
    pass


@app.get(
    '/experiments/{experiment_id}/snapshot/{snapshot_id}',
    response_model=SnapshotStatus,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Experiments'],
)
def get_experiment_snapshot(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    snapshot_id: UUID = ...,
) -> Union[SnapshotStatus, Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/vlan/{source_lan}/connect',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def connect_experiment_vlan(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    source_lan: str = ...,
    target_id: PortalID = ...,
    target_lan: str = ...,
) -> Optional[Error]:
    pass


@app.post(
    '/experiments/{experiment_id}/vlan/{source_lan}/disconnect',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Experiments'],
)
def disconnect_experiment_vlan(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    experiment_id: PortalID = ...,
    source_lan: str = ...,
) -> Optional[Error]:
    pass


@app.get(
    '/profiles',
    response_model=ProfileList,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
    },
    tags=['Profiles'],
)
def list_profiles(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    x__api__elaborate: Optional[str] = Header(None, alias='X-Api-Elaborate'),
    profile_id: Optional[UUID] = None,
    creator: Optional[str] = None,
    project: Optional[str] = None,
    profile_name: Optional[str] = None,
) -> Union[ProfileList, Error]:
    pass


@app.post(
    '/profiles',
    response_model=None,
    responses={
        '201': {'model': Profile},
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '429': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Profiles'],
)
def create_profile(
    x__api__token: str = Header(..., alias='X-Api-Token'), body: ProfileCreate = None
) -> Optional[Union[Profile, Error]]:
    pass


@app.get(
    '/profiles/{profile_id}',
    response_model=Profile,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Profiles'],
)
def get_profile(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    x__api__elaborate: Optional[str] = Header(None, alias='X-Api-Elaborate'),
    profile_id: PortalID = ...,
) -> Union[Profile, Error]:
    pass


@app.patch(
    '/profiles/{profile_id}',
    response_model=Profile,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Profiles'],
)
def modify_profile(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    profile_id: PortalID = ...,
    body: ProfileModify = None,
) -> Union[Profile, Error]:
    pass


@app.put(
    '/profiles/{profile_id}',
    response_model=Profile,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Profiles'],
)
def update_profile(
    x__api__token: str = Header(..., alias='X-Api-Token'), profile_id: PortalID = ...
) -> Union[Profile, Error]:
    pass


@app.delete(
    '/profiles/{profile_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Profiles'],
)
def delete_profile(
    x__api__token: str = Header(..., alias='X-Api-Token'), profile_id: PortalID = ...
) -> Optional[Error]:
    pass


@app.get(
    '/profiles/{profile_id}/versions/{version_id}',
    response_model=Profile,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Profiles'],
)
def get_profile_version(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    profile_id: PortalID = ...,
    version_id: UUID = ...,
) -> Union[Profile, Error]:
    pass


@app.delete(
    '/profiles/{profile_id}/versions/{version_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Profiles'],
)
def delete_profile_version(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    profile_id: PortalID = ...,
    version_id: UUID = ...,
) -> Optional[Error]:
    pass


@app.get(
    '/resgroups',
    response_model=ResGroupList,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
    },
    tags=['Resgroups'],
)
def list_resgroups(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    x__api__elaborate: Optional[str] = Header(None, alias='X-Api-Elaborate'),
    resgroup_id: Optional[UUID] = None,
    creator: Optional[str] = None,
    project: Optional[str] = None,
) -> Union[ResGroupList, Error]:
    pass


@app.post(
    '/resgroups',
    response_model=None,
    responses={
        '201': {'model': ResGroup},
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '406': {'model': ResGroupError},
        '500': {'model': Error},
    },
    tags=['Resgroups'],
)
def create_resgroup(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    duration: Optional[int] = None,
    noautoapprove: Optional[bool] = False,
    body: ResGroup = None,
) -> Optional[Union[ResGroup, Error, ResGroupError]]:
    pass


@app.post(
    '/resgroups/search',
    response_model=ResGroupSearchResult,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '406': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Resgroups'],
)
def search_resgroup(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    duration: int = ...,
    body: ResGroupSearch = None,
) -> Union[ResGroupSearchResult, Error]:
    pass


@app.get(
    '/resgroups/{resgroup_id}',
    response_model=ResGroup,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Resgroups'],
)
def get_resgroup(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    x__api__elaborate: Optional[str] = Header(None, alias='X-Api-Elaborate'),
    resgroup_id: UUID = ...,
) -> Union[ResGroup, Error]:
    pass


@app.put(
    '/resgroups/{resgroup_id}',
    response_model=ResGroup,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Resgroups'],
)
def modify_resgroup(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    resgroup_id: UUID = ...,
    duration: Optional[int] = None,
    noautoapprove: Optional[bool] = False,
    body: ResGroup = None,
) -> Union[ResGroup, Error]:
    pass


@app.delete(
    '/resgroups/{resgroup_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Resgroups'],
)
def delete_resgroup(
    x__api__token: str = Header(..., alias='X-Api-Token'), resgroup_id: UUID = ...
) -> Optional[Error]:
    pass


@app.post(
    '/resgroups/{resgroup_id}/reservations',
    response_model=None,
    responses={
        '201': {'model': ResGroup},
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
        '406': {'model': ResGroupError},
        '500': {'model': Error},
    },
    tags=['Resgroups'],
)
def add_resgroup_reservation(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    resgroup_id: UUID = ...,
    noautoapprove: Optional[bool] = False,
    body: ResGroupReservation = None,
) -> Optional[Union[ResGroup, Error, ResGroupError]]:
    pass


@app.delete(
    '/resgroups/{resgroup_id}/reservations/{reservation_id}',
    response_model=None,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '404': {'model': Error},
    },
    tags=['Resgroups'],
)
def delete_resgroup_reservation(
    x__api__token: str = Header(..., alias='X-Api-Token'),
    resgroup_id: UUID = ...,
    reservation_id: UUID = ...,
) -> Optional[Error]:
    pass


@app.get(
    '/tokens/this',
    response_model=Token,
    responses={
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Token'],
)
def get_token_this(
    x__api__token: str = Header(..., alias='X-Api-Token')
) -> Union[Token, Error]:
    pass


@app.put(
    '/tokens/this/refresh',
    response_model=None,
    responses={
        '201': {'model': Token},
        '400': {'model': Error},
        '401': {'model': Error},
        '403': {'model': Error},
        '500': {'model': Error},
    },
    tags=['Token'],
)
def refresh_token_this(
    x__api__token: str = Header(..., alias='X-Api-Token')
) -> Optional[Union[Token, Error]]:
    pass


@app.get('/version', response_model=Version, tags=['Version'])
def get_version() -> Version:
    pass